back-django/requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.9
django-cors-headers==4.8.0
djangorestframework==3.16.1
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
setuptools==78.1.1
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
pylint==3.3.8
djangorestframework-simplejwt==5.5.0
